Add GenericMaterial class and update material imports
- Introduced GenericMaterial class for base material functionality.
- Updated __all__ in __init__.py to include GenericMaterial.
- Modified Stone class to inherit from GenericMaterial.
- Enhanced Solver class with data handling methods.
