+++ src/compas_dem/material/__init__.py
@@ -1,3 +1,4 @@
+from .generic import GenericMaterial
 from .stone import Stone
 
-__all__ = ["Stone"]
+__all__ = ["GenericMaterial", "Stone"]
